Fix tests, remove duplicate tests
