+++ BTCPayServer.Tests/PayJoinTests.cs
@@ -250,11 +250,11 @@ namespace BTCPayServer.Tests
             await s.StartAsync();
             s.RegisterNewUser(true);
             var receiver = s.CreateNewStore();
-            s.GenerateWallet("BTC", "", true, true, ScriptPubKeyType.Segwit);
+            s.GenerateWallet("BTC", "", true, true);
             var receiverWalletId = new WalletId(receiver.storeId, "BTC");
 
             var sender = s.CreateNewStore();
-            s.GenerateWallet("BTC", "", true, true, ScriptPubKeyType.Segwit);
+            s.GenerateWallet("BTC", "", true, true);
             var senderWalletId = new WalletId(sender.storeId, "BTC");
 
             await s.Server.ExplorerNode.GenerateAsync(1);
@@ -263,8 +263,7 @@ namespace BTCPayServer.Tests
 
             var invoiceId = s.CreateInvoice(receiver.storeId, null, "BTC");
             s.GoToInvoiceCheckout(invoiceId);
-            var bip21 = s.Driver.FindElement(By.ClassName("payment__details__instruction__open-wallet__btn"))
-                .GetAttribute("href");
+            var bip21 = s.Driver.WaitForElement(By.Id("PayInWallet")).GetAttribute("href");
             Assert.Contains($"{PayjoinClient.BIP21EndpointKey}=", bip21);
             s.GoToWallet(senderWalletId, WalletsNavPages.Send);
             s.Driver.FindElement(By.Id("bip21parse")).Click();
@@ -310,8 +309,7 @@ namespace BTCPayServer.Tests
                 //payjoin is enabled by default.
                 var invoiceId = s.CreateInvoice(receiver.storeId);
                 s.GoToInvoiceCheckout(invoiceId);
-                var bip21 = s.Driver.WaitForElement(By.ClassName("payment__details__instruction__open-wallet__btn"))
-                    .GetAttribute("href");
+                var bip21 = s.Driver.WaitForElement(By.Id("PayInWallet")).GetAttribute("href");
                 Assert.Contains($"{PayjoinClient.BIP21EndpointKey}=", bip21);
 
                 s.GoToStore(receiver.storeId);
@@ -326,8 +324,7 @@ namespace BTCPayServer.Tests
 
                 invoiceId = s.CreateInvoice(receiver.storeId);
                 s.GoToInvoiceCheckout(invoiceId);
-                bip21 = s.Driver.WaitForElement(By.ClassName("payment__details__instruction__open-wallet__btn"))
-                    .GetAttribute("href");
+                bip21 = s.Driver.WaitForElement(By.Id("PayInWallet")).GetAttribute("href");
                 Assert.Contains($"{PayjoinClient.BIP21EndpointKey}=", bip21);
 
                 s.GoToWallet(senderWalletId, WalletsNavPages.Send);
@@ -360,8 +357,7 @@ namespace BTCPayServer.Tests
                 //let's do it all again, except now the receiver has funds and is able to payjoin
                 invoiceId = s.CreateInvoice();
                 s.GoToInvoiceCheckout(invoiceId);
-                bip21 = s.Driver.WaitForElement(By.ClassName("payment__details__instruction__open-wallet__btn"))
-                    .GetAttribute("href");
+                bip21 = s.Driver.WaitForElement(By.Id("PayInWallet")).GetAttribute("href");
                 Assert.Contains($"{PayjoinClient.BIP21EndpointKey}", bip21);
 
                 s.GoToWallet(senderWalletId, WalletsNavPages.Send);
@@ -378,7 +374,7 @@ namespace BTCPayServer.Tests
                     s.Driver.FindElement(By.CssSelector("button[value=payjoin]")).Click();
                     return Task.CompletedTask;
                 });
-                s.FindAlertMessage(StatusMessageModel.StatusSeverity.Success);
+                s.FindAlertMessage();
                 var handler = s.Server.PayTester.GetService<PaymentMethodHandlerDictionary>().GetBitcoinHandler("BTC");
                 await TestUtils.EventuallyAsync(async () =>
                 {
